+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add way to write an event comment when making event. Show in UI (both viewing and when taking actions).
+
+
+
 ## [0.1.0] - 2020-05-29
 
 First release of new project